Fix Gradle build deprecations and warnings

+++ app/src/main/kotlin/com/music/chorus/utils/ComposeToImage.kt
@@ -112,7 +112,7 @@ object ComposeToImage {
                             
                             
                             val overlayPaint = Paint().apply {
-                                color = 0x4D000000.toInt() 
+                                color = 0x4D000000 
                             }
                             canvas.drawRect(blurRect, overlayPaint)
                         }
